+++ src/test/java/GsonHandlerTest.java
@@ -1,14 +1,26 @@
 import org.junit.jupiter.api.Test;
 
+import java.io.File;
+
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 
-class FileHandlerTest {
-    FileHandler handler = new FileHandler();
+/**
+ * Simple test to test GsonHandler by writing values to a file and reading values from the file,
+ * testing FileReader.java by extension.
+ */
+class GsonHandlerTest {
+    GsonHandler handler = new GsonHandler();
     ToDoItem item = new ToDoItem();
     ToDoList test = new ToDoList();
+    File file = new File("save.txt");
 
+    /**
+     * Creates a simple list of "Active item: " + the current iteration value
+     * for values from 0-10 and then adds them to a list and notifies subscribers of
+     * changes.
+     */
     void createListValues() {
         for (int i = 0; i < 10; i++) {
             item.setDescription("Active item: " + i);
@@ -20,6 +32,10 @@ class FileHandlerTest {
         }
     }
 
+    /**
+     * Writes precreated values to the Gson and save file
+     * and then asserts that the written values contain the items written to it.
+     */
     @Test
     void writeToFile() {
         createListValues();
@@ -28,6 +44,12 @@ class FileHandlerTest {
         assertTrue(handler.getfOutput().contains(item.getDescription()));
     }
 
+    /**
+     * Tests both readFromFile() in GsonHandler along with FileReader by extension
+     * by writing to file, then reading from the file and testing to make sure the items
+     * in the original list match the ouput list and the list written to file.
+     * and then deletes it to ensure that tested values are not in the actual list.
+     */
     @Test
     void readFromFile() {
         handler.readFromFile();
@@ -35,5 +57,6 @@ class FileHandlerTest {
         handler.readFromFile();
 //        System.out.println(handler.getList().getItems());
         assertEquals(handler.getList().getItems(), test.getItems());
+        file.delete();
     }
 }