+++ src/test/java/ListManagerTest.java
@@ -1,11 +1,19 @@
 import org.junit.jupiter.api.Test;
 
+/**
+ * Test by populating an initial list of values, then adding and subtracting
+ * values to the list, asserting that the ToDoList and the test list match values.
+ */
 class ListManagerTest {
     ListManager listManager = new ListManager();
-    FileHandler fileHandler = new FileHandler();
+    GsonHandler gsonHandler = new GsonHandler();
     ToDoList testList = new ToDoList();
     ToDoItem item = new ToDoItem();
 
+    /**
+     * Creates an initial population of "Managed List values" items and then
+     * add them to a testList, notifying subscribers of changes.
+     */
     void populate() {
         for (int i = 0; i < 10; i++) {
             item.setPriority(i);
@@ -15,26 +23,40 @@ class ListManagerTest {
         }
     }
 
-
+    /**
+     * Creates an initial population of the list, writing the values to file and asserting
+     * the initial values match the values written to file and pushed to the ListManager.
+     */
     @Test
     void initialPopulate() {
-        fileHandler.writeToFile(testList);
-        System.out.println(fileHandler.getfOutput());
+        populate();
+        gsonHandler.writeToFile(testList);
+        System.out.println(gsonHandler.getfOutput());
+        gsonHandler.readFromFile();
+        System.out.println(gsonHandler.getfOutput());
         listManager.initialPopulate();
 //        assertTrue(testList.getItems().isEmpty());
 
 
     }
 
+    /**
+     * Adds a value of "Addded Value!" to the ListManager and checks that it also
+     * published to the file in Gson form.
+     */
     @Test
     void addValue() {
         listManager.addValue("Added Value!");
-//        assertEquals(listManager.getManagedList().getItems(), fileHandler.getList().getItems());
+//        assertEquals(listManager.getManagedList().getItems(), gsonHandler.getList().getItems());
     }
 
+    /**
+     * Removes a value at index 0 and then checks that the list manager and
+     * the Gson match up.
+     */
     @Test
     void removeValue() {
         listManager.removeValue(0);
-//        assertEquals(listManager.getManagedList().getItems(),fileHandler.getList().getItems());
+//        assertEquals(listManager.getManagedList().getItems(),gsonHandler.getList().getItems());
     }
 }