+++ src/test/java/PublisherTest.java
@@ -3,6 +3,10 @@ import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
+/**
+ * Tests that Observers can be added to a list of subscribers and can be notified and
+ * unsubscribed from the list.
+ */
 class PublisherTest {
     Observer testObserver = () -> {
 
@@ -15,6 +19,10 @@ class PublisherTest {
         }
     };
 
+    /**
+     * Tests that an empty testObserver is in the list of observers
+     * and subscribed for updates.
+     */
     @Test
     void subscribe() {
         publisher.subscribe(testObserver);
@@ -23,6 +31,10 @@ class PublisherTest {
 
     }
 
+    /**
+     * Removes the testObserver from the list of observers and tests
+     * that it was correctly removed.
+     */
     @Test
     void unsubscribe() {
         subscribe();
@@ -30,11 +42,15 @@ class PublisherTest {
         assertTrue(publisher.subscribers.isEmpty());
     }
 
+    /**
+     * Subscribes an empty testObserver to a notify list and then asserts that
+     * the publisher publishes changes to the testObserver.
+     */
     @Test
     void notifySubscribers() {
         subscribe();
         publisher.notifySubscribers();
-        System.out.println(publisher.subscribers.get(0));
+//        System.out.println(publisher.subscribers.get(0));
         assertTrue(publisher.subscribers.contains(testObserver));
     }
 }