+++ CHANGELOG.md
@@ -16,11 +16,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Made crucial changes to UML Diagram
 - Changed permissions in gradlew
 - Changed from --scan to just build in travis ci
-- Changed FileHandlerTest.java now that the structure is updated.
+- Changed GsonHandlerTest.java now that the structure is updated.
 ### Added
-- Creating simple tests for ToDoList, readFromFile, FileHandler and Publisher.
+- Creating simple tests for ToDoList, readFromFile, GsonHandler and Publisher.
 - Created Observer, Publisher, ToDoItem, ToDoList, and ToDoListGUI to exhibit proper Observer Class behavior.
-- Wrote readFromFile and FileHandler to use GSon and newly written classes.
+- Wrote readFromFile and GsonHandler to use GSon and newly written classes.
 - Committed JavaDocs
 - Added UML Design Image
 - Added UI Design Image